# api/openapi/management-v1.yaml
openapi: 3.1.0
info:
  title: MarketplaceServer Management API
  version: 1.0.0
  license:
    name: MIT
    identifier: MIT
  description: |
    Authoritative wire contract for currently implemented management operations
    under `/api/v1`.

    This document contains deployed login, health, current-user personal access
    token, Plugin lifecycle, and Plugin Version operations. Future Marketplace,
    team, and distribution credential management paths remain in
    `management-v1-design.yaml` and are not evidence of runtime availability.

    Git Smart HTTP, distribution endpoints, frontend routes, and the external
    Claude Code `marketplace.json` schema are outside this contract.
  x-document-status: approved
  x-implementation-conformance: implemented
...
servers:
  - url: /
    description: Same-origin MarketplaceServer deployment
...
paths:
...
  /api/v1/namespaces/{namespace}/plugins:
    parameters:
      - $ref: "#/components/parameters/Namespace"
    get:
      operationId: listPlugins
      tags: [Plugins]
      summary: List readable Plugins in a namespace
      security:
        - BearerJWT: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Stable namespace-scoped Plugin page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PluginPageSuccess"
        "401":
          $ref: "#/components/responses/BearerUnauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createPlugin
      tags: [Plugins]
      summary: Create a Plugin and its hidden shared-ID repository
      description: |
        The lowercase kebab-case name is the immutable Plugin slug and Git path.
        Visibility defaults to public. A 201 is synchronous and is returned only
        after the same-UUID hidden bare repository is initialized with symbolic
        HEAD pointing to main and both aggregate SQL rows are ready. SQL and Git
        remain separate atomic boundaries with explicit compensation.
      security:
        - BearerJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePluginRequest"
      responses:
        "201":
          description: Draft Plugin and hidden Repository aggregate created and ready.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PluginSuccess"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/BearerUnauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/namespaces/{namespace}/plugins/{plugin}:
    parameters:
      - $ref: "#/components/parameters/Namespace"
      - $ref: "#/components/parameters/PluginSlug"
    get:
      operationId: getPlugin
      tags: [Plugins]
      summary: Get Plugin metadata
      security:
        - {}
        - BearerJWT: []
      responses:
        "200":
          description: Exact public active/archived metadata is anonymous; draft/private require authorization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PluginSuccess"
        "401":
          $ref: "#/components/responses/BearerUnauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/namespaces/{namespace}/plugins/{plugin}/versions:
    parameters:
      - $ref: "#/components/parameters/Namespace"
      - $ref: "#/components/parameters/PluginSlug"
    get:
      operationId: listPluginVersions
      tags: [Plugin versions]
      summary: List available logical tag Versions
      security:
        - BearerJWT: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Stable page excluding deleted tombstones.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PluginVersionPageSuccess"
        "401":
          $ref: "#/components/responses/BearerUnauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/namespaces/{namespace}/plugins/{plugin}/versions/{tag}:
    parameters:
      - $ref: "#/components/parameters/Namespace"
      - $ref: "#/components/parameters/PluginSlug"
      - $ref: "#/components/parameters/TagName"
    get:
      operationId: getPluginVersion
      tags: [Plugin versions]
      summary: Get an exact logical Version
      security:
        - {}
        - BearerJWT: []
      responses:
        "200":
          description: Available Version metadata under current Plugin visibility.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PluginVersionSuccess"
        "401":
          $ref: "#/components/responses/BearerUnauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "410":
          $ref: "#/components/responses/Gone"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    BearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Stateless management JWT signed with HS256. Issued tokens contain exactly
        canonical username, iat, and exp claims and have a fixed 30-day lifetime.
        Account passwords and PATs are not accepted as management Bearer tokens.
  parameters:
    Page:
      name: page
      in: query
      required: false
      description: One-based page number; defaults to 1.
      schema:
        type: integer
        minimum: 1
        default: 1
    Size:
      name: size
      in: query
      required: false
      description: Page size; defaults to 20 and must not exceed 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    Namespace:
      name: namespace
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Slug"
    PluginSlug:
      name: plugin
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Slug"
    TagName:
      name: tag
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/CanonicalVersionTag"
  headers:
    RequestId:
      description: Framework-generated ULID matching the error body's requestId.
      schema:
        type: string
        pattern: "^[0-9A-HJKMNP-TV-Z]{26}$"
    BearerChallenge:
      description: Management Bearer authentication challenge.
      schema:
        type: string
        const: 'Bearer realm="MarketplaceServer Management"'
...
  responses:
    BadRequest:
      description: The body is missing, unreadable, malformed, oversized, top-level null, or contains a trailing JSON value.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: bad_request
            message: request body is invalid
            requestId: 01J51X8R6M4B9D5VYN2Q3H7KCT
...
    BearerUnauthenticated:
      description: The management Bearer JWT is missing or invalid.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
        WWW-Authenticate:
          $ref: "#/components/headers/BearerChallenge"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: unauthenticated
            message: authentication is required
            requestId: 01J51X8R6M4B9D5VYN2Q3H7KCT
...
    Forbidden:
      description: The authenticated principal is not authorized for this action and resource.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: forbidden
            message: permission denied
            requestId: 01J51X8R6M4B9D5VYN2Q3H7KCT
...
    NotFound:
      description: The scoped resource was not found or is intentionally not disclosed.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: The resource state, version, or unique identity conflicts with the command.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Gone:
      description: The authorized exact logical Version is retained only as a deleted tombstone; no deleted metadata is disclosed.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationFailed:
      description: Query syntax or token input violates the documented constraints.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: validation_failed
            message: request is invalid
            requestId: 01J51X8R6M4B9D5VYN2Q3H7KCT
    InternalError:
      description: Internal failure without implementation detail disclosure.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: internal_error
            message: internal server error
            requestId: 01J51X8R6M4B9D5VYN2Q3H7KCT
  schemas:
...
    Timestamp:
      type: string
      format: date-time
      description: RFC 3339 timestamp.
    Slug:
      type: string
      minLength: 1
      maxLength: 63
      pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
    CommitSHA:
      type: string
      pattern: "^(?:[0-9a-f]{40}|[0-9a-f]{64})$"
      description: Full peeled commit object ID in the repository object format.
    CanonicalVersionTag:
      type: string
      pattern: "^v(?:0|[1-9][0-9]*)\\.(?:0|[1-9][0-9]*)\\.(?:0|[1-9][0-9]*)(?:-[0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*)?(?:\\+[0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*)?$"
      description: Canonical v-prefixed SemVer tag.
    Error:
      type: object
      additionalProperties: false
      required: [code, message, requestId]
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
          pattern: "^[0-9A-HJKMNP-TV-Z]{26}$"
        details:
          type: string
          description: Optional safe text; clients must branch on code, not this value.
...
    PluginStatus:
      type: string
      enum: [draft, active, archived]
      description: First successful Version publish activates draft; archive is explicitly reversible.
    PluginVisibility:
      type: string
      enum: [public, private]
    RepositoryStatus:
      type: string
      enum: [ready, readOnly, error]
      description: Safe operational state of the mandatory hidden same-ID Repository.
    Plugin:
      type: object
      required: [namespace, name, status, visibility, repositoryStatus, cloneUrl, defaultVersion, createdAt, updatedAt]
      properties:
        namespace:
          $ref: "#/components/schemas/Slug"
        name:
          $ref: "#/components/schemas/Slug"
        status:
          $ref: "#/components/schemas/PluginStatus"
        visibility:
          $ref: "#/components/schemas/PluginVisibility"
        repositoryStatus:
          $ref: "#/components/schemas/RepositoryStatus"
        cloneUrl:
          type: string
          pattern: "^/git/[a-z0-9]+(?:-[a-z0-9]+)*/[a-z0-9]+(?:-[a-z0-9]+)*\\.git$"
          description: Same-origin relative development Git URL; no Repository ID or storage detail.
        defaultVersion:
          oneOf:
            - $ref: "#/components/schemas/CanonicalVersionTag"
            - type: "null"
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        updatedAt:
          $ref: "#/components/schemas/Timestamp"
    CreatePluginRequest:
      type: object
      required: [name]
      properties:
        name:
          $ref: "#/components/schemas/Slug"
        visibility:
          $ref: "#/components/schemas/PluginVisibility"
...
    PluginSuccess:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Plugin"
    PluginVersionStatus:
      type: string
      const: available
      description: Deleted logical Versions are omitted from lists and return a 410 error for an authorized exact query.
    PluginVersion:
      type: object
      required: [tag, status, commitSha, publishedAt, updatedAt]
      properties:
        tag:
          $ref: "#/components/schemas/CanonicalVersionTag"
        status:
          $ref: "#/components/schemas/PluginVersionStatus"
        commitSha:
          $ref: "#/components/schemas/CommitSHA"
        publishedAt:
          $ref: "#/components/schemas/Timestamp"
        updatedAt:
          $ref: "#/components/schemas/Timestamp"
...
    PluginVersionSuccess:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/PluginVersion"
    PageMeta:
      type: object
      additionalProperties: false
      required: [page, size, total]
      properties:
        page:
          type: integer
          minimum: 1
        size:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          format: int64
          minimum: 0
...
    PluginPageData:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Plugin"
    PluginVersionPageData:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/PluginVersion"
...
    PluginPageSuccess:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/PluginPageData"
    PluginVersionPageSuccess:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/PluginVersionPageData"